Move animation to Character class
